Fixing lots of minor things as requested in PR
