+++ Frends.Excel.ConvertToCSV/Frends.Excel.ConvertToCSV/FrendsTaskMetadata.json
@@ -1,0 +1,7 @@
+﻿{
+	"Tasks": [
+		{
+			"TaskMethod": "Frends.Excel.ConvertToCSV.Excel.ConvertToCSV"
+		}
+	]
+}
