+++ Frends.Excel.ConvertToXML/Frends.Excel.ConvertToXML/Frends.Excel.ConvertToXML.csproj
@@ -26,5 +26,9 @@
   <ItemGroup>
     <PackageReference Include="ExcelDataReader.DataSet" Version="3.6.0" />
     <PackageReference Include="ExcelDataReader" Version="3.6.0" />
+    <PackageReference Include="SonarAnalyzer.CSharp" Version="8.36.0.43782">
+      <PrivateAssets>all</PrivateAssets>
+      <IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
+    </PackageReference>
   </ItemGroup>
 </Project>
