+++ Frends.Excel.ConvertToXML/Frends.Excel.ConvertToXML/FrendsTaskMetadata.json
@@ -1,0 +1,7 @@
+﻿{
+	"Tasks": [
+		{
+			"TaskMethod": "Frends.Excel.ConvertToXML.Excel.ConvertToXML"
+		}
+	]
+}
